+++ services/s3/src/it/java/software/amazon/awssdk/services/s3/crossregion/S3CrossRegionAsyncIntegrationTestBase.java
@@ -15,18 +15,13 @@
 
 package software.amazon.awssdk.services.s3.crossregion;
 
-import static software.amazon.awssdk.testutils.service.S3BucketUtils.temporaryBucketName;
-
 import java.util.ArrayList;
-import java.util.Iterator;
 import java.util.List;
-import org.junit.jupiter.api.AfterAll;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.BeforeEach;
+import java.util.concurrent.CompletableFuture;
 import software.amazon.awssdk.core.ResponseBytes;
-import software.amazon.awssdk.core.sync.RequestBody;
-import software.amazon.awssdk.core.sync.ResponseTransformer;
-import software.amazon.awssdk.services.s3.S3Client;
+import software.amazon.awssdk.core.async.AsyncRequestBody;
+import software.amazon.awssdk.core.async.AsyncResponseTransformer;
+import software.amazon.awssdk.services.s3.S3AsyncClient;
 import software.amazon.awssdk.services.s3.model.DeleteObjectRequest;
 import software.amazon.awssdk.services.s3.model.DeleteObjectResponse;
 import software.amazon.awssdk.services.s3.model.DeleteObjectsRequest;
@@ -36,72 +31,47 @@ import software.amazon.awssdk.services.s3.model.GetObjectResponse;
 import software.amazon.awssdk.services.s3.model.HeadBucketRequest;
 import software.amazon.awssdk.services.s3.model.HeadBucketResponse;
 import software.amazon.awssdk.services.s3.model.ListObjectsV2Request;
-import software.amazon.awssdk.services.s3.model.ListObjectsV2Response;
 import software.amazon.awssdk.services.s3.model.PutObjectRequest;
 import software.amazon.awssdk.services.s3.model.PutObjectResponse;
 import software.amazon.awssdk.services.s3.model.S3Object;
+import software.amazon.awssdk.services.s3.paginators.ListObjectsV2Publisher;
 
-public class S3SyncCrossRegionIntegrationTest extends S3CrossRegionIntegrationTestBase {
-    private static final String BUCKET = temporaryBucketName(S3SyncCrossRegionIntegrationTest.class);
-    private S3Client crossRegionS3Client;
-
-    @BeforeAll
-    static void setUpClass() {
-        s3 = s3ClientBuilder().build();
-        createBucket(BUCKET);
-    }
-
-    @AfterAll
-    static void clearClass() {
-        deleteBucketAndAllContents(BUCKET);
-    }
-
-    @BeforeEach
-    public void initialize() {
-        crossRegionS3Client = S3Client.builder()
-                                      .region(CROSS_REGION)
-                                      .serviceConfiguration(s -> s.crossRegionAccessEnabled(true))
-                                      .build();
-    }
+public abstract class S3CrossRegionAsyncIntegrationTestBase extends S3CrossRegionIntegrationTestBase{
+    protected S3AsyncClient crossRegionS3Client;
 
     @Override
     protected List<S3Object> paginatedAPICall(ListObjectsV2Request listObjectsV2Request) {
-        List<S3Object> resultS3Object = new ArrayList<>();
-        Iterator<ListObjectsV2Response> v2ResponseIterator =
-            crossRegionS3Client.listObjectsV2Paginator(listObjectsV2Request).iterator();
-        while (v2ResponseIterator.hasNext()) {
-            v2ResponseIterator.next().contents().forEach(a -> resultS3Object.add(a));
-        }
-        return resultS3Object;
+        List<S3Object> resultObjects = new ArrayList<>();
+        ListObjectsV2Publisher publisher = crossRegionS3Client.listObjectsV2Paginator(listObjectsV2Request);
+        CompletableFuture<Void> subscribe = publisher.subscribe(response -> {
+            response.contents().forEach(a -> resultObjects.add(a));
+        });
+        subscribe.join();
+        return resultObjects;
     }
 
     @Override
     protected DeleteObjectsResponse postObjectAPICall(DeleteObjectsRequest deleteObjectsRequest) {
-        return crossRegionS3Client.deleteObjects(deleteObjectsRequest);
+        return crossRegionS3Client.deleteObjects(deleteObjectsRequest).join();
     }
 
     @Override
     protected HeadBucketResponse headAPICall(HeadBucketRequest headBucketRequest) {
-        return crossRegionS3Client.headBucket(headBucketRequest);
+        return crossRegionS3Client.headBucket(headBucketRequest).join();
     }
 
     @Override
     protected DeleteObjectResponse deleteObjectAPICall(DeleteObjectRequest deleteObjectRequest) {
-        return crossRegionS3Client.deleteObject(deleteObjectRequest);
+        return crossRegionS3Client.deleteObject(deleteObjectRequest).join();
     }
 
     @Override
     protected PutObjectResponse putAPICall(PutObjectRequest putObjectRequest, String testString) {
-        return crossRegionS3Client.putObject(putObjectRequest, RequestBody.fromString(testString));
+        return crossRegionS3Client.putObject(putObjectRequest, AsyncRequestBody.fromString(testString)).join();
     }
 
     @Override
     protected ResponseBytes<GetObjectResponse> getAPICall(GetObjectRequest getObjectRequest) {
-        return crossRegionS3Client.getObject(getObjectRequest, ResponseTransformer.toBytes());
-    }
-
-    @Override
-    protected String bucketName() {
-        return BUCKET;
+        return crossRegionS3Client.getObject(getObjectRequest, AsyncResponseTransformer.toBytes()).join();
     }
 }
